Debut de raccordement WatchdogdAdmin au serveur via ZMQ tcp
git-svn-id: https://svn.abls-habitat.fr/repo/Watchdog@3908 9f22af27-4b39-4cfe-9f85-4bed9fef4693

+++ trunk/Watchdogd/Include/Zmq.h
@@ -44,14 +44,14 @@
   };
 
  struct ZMQ_TARGET
-  { guint tag;
+  { gint8 tag;
     gchar instance[12];
     gchar thread[12];
   };
 
  struct ZMQ_SET_BIT
-  { gint  type;
-    gint  num;
+  { gint8 type;
+    gint8 num;
     gchar acronyme [ NBR_CARAC_ACRONYME_MNEMONIQUE_UTF8+1 ];
     gchar dls_tech_id [ NBR_CARAC_PLUGIN_DLS_TECHID ];
   };
